#35 / Refactor createshortgame package
