+++ app/src/main/java/sopt/uni/presentation/shortgame/createshortgame/CreateShortGameActivity.kt
@@ -1,4 +1,4 @@
-package sopt.uni.presentation.shortgame
+package sopt.uni.presentation.shortgame.createshortgame
 
 import android.content.Context
 import android.graphics.Rect
