+++ app/src/main/java/sopt/uni/presentation/shortgame/createshortgame/CreateShortGameViewModel.kt
@@ -1,4 +1,4 @@
-package sopt.uni.presentation.shortgame
+package sopt.uni.presentation.shortgame.createshortgame
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
